Make duke package
