+++ src/main/java/duke/Cortex.java
@@ -1,4 +1,4 @@
-import java.util.ArrayList;
+package duke;
 
 public class Cortex {
     public TaskList list;
