+++ src/main/java/duke/Ui.java
@@ -1,3 +1,4 @@
+package duke;
 import java.util.ArrayList;
 import java.util.Scanner;
 public class Ui {
